+++ src/com/v5kf/mcss/utils/UITools.java
@@ -374,9 +374,9 @@ public class UITools {
 	 *            （DisplayMetrics类中属性density）
 	 * @return
 	 */
-	public static int dip2px(Context context, float dipValue) {
+	public static float dip2px(Context context, float dipValue) {
 		final float scale = context.getResources().getDisplayMetrics().density;
-		return (int) (dipValue * scale + 0.5f);
+		return (dipValue * scale + 0.5f);
 	}
 
 	/**
@@ -387,7 +387,7 @@ public class UITools {
 	 *            （DisplayMetrics类中属性scaledDensity）
 	 * @return
 	 */
-	public static int px2sp(Context context, float pxValue) {
+	public static float px2sp(Context context, float pxValue) {
 		final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
 		return (int) (pxValue / fontScale + 0.5f);
 	}
